src/grpc_services/requirements.txt
# gRPC and related tools
grpcio==1.51.1
grpcio-tools==1.51.1
grpcio-health-checking==1.51.1

# Resilience and monitoring
tenacity>=8.2.2
prometheus-client>=0.16.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.3  # For PostgreSQL support